Library update

+++ main/misc.c
@@ -419,7 +419,7 @@ static bool is_valid_int(const char *str)
    // Check for non-digit chars in the rest of the stirng.
    while (*str)
    {
-      if (!isdigit(*str))
+      if (!isdigit((int)*str))
          return false;
       else
          ++str;
